修改节点形状

+++ src/main/java/com/example/neo4jKG/Dao/NeoEntityRepository.java
@@ -16,7 +16,7 @@ public interface NeoEntityRepository extends Neo4jRepository<NeoEntity, Long> {
             "n.category= :#{#category}, n.symbolSize = :#{#symbolSize}, n.centerX = :#{#centerX}, n.centerY = :#{#centerY}," +
             "n.symbol= :#{#symbol} RETURN n")
     NeoEntity updateByEntity(@Param(value = "id") Long id, @Param(value = "name") String name, @Param(value = "x") double x,
-                             @Param(value = "y") double y, @Param(value = "des") String des, @Param(value = "category") int category,
+                             @Param(value = "y") double y, @Param(value = "des") String des, @Param(value = "category") long category,
                              @Param(value = "symbolSize") int symbolSize, @Param(value = "centerX") double centerX,
                              @Param(value = "centerY") double centerY,@Param(value = "symbol") String symbol);
